Update arduino-BerryIMU.ino
Changing the 16 bits of raw reading results to type 'short' (int16), rather than type 'int' (int32). Casting a 2-byte value to the 4-byte 'int' fails to propagate the sign bit, which means all of the *Raw values show up as unsigned 16-bit integers (0 to +65535), instead of signed values (-32768 to +32767).

The most noticeable result is that the angle of the heading always shows up in the first quadrant, i.e. between 0 and 90 degrees.

Casting the 2-byte composite value to a 'short' yields the desired sign value, and when it is assigned to an 'int' variable, the sign is preserved.